Pipeline for parent-independent genotyping-by-sequencing in potato.
a, High-quality heterozygous SNPs in the F1 clone were extracted. b, The genotype of these SNPs in each F2 individual was identified from low-coverage sequencing data. The breakpoints between homozygous and heterozygous regions were deduced on the basis of the ratio of heterozygous SNPs to all SNPs in each window (window size, 1 Mb; step, 100 kb), and all SNPs in homozygous regions were extracted. c, The longest homozygous region or two overlapping homozygous regions that covered the whole chromosome were chosen as the reference fragments. All homozygous fragments were compared with the reference fragments and divided into two groups based on similarity to the reference fragment. d, Within each group, all SNPs with consensus at each locus were combined to construct the haplotype. e, The weighted value of genotype ‘b’ in each window was used to estimate the breakpoints of recombination. f, The genotype of a bin was defined on the basis of the weighted value of genotype ‘b’ in each bin. Blue, yellow, and gray indicate genotype ‘a’, ‘b’, and ‘h’, respectively. The same procedure was carried out for all 12 chromosomes.